backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djoser==2.2.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pyflakes==3.1.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
python-dotenv==1.0.0
